backend/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
